backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.7
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose==3.4.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
redis==5.2.1
prometheus-client==0.21.1
python-json-logger==2.0.7
orjson==3.10.16
pytest==8.3.5
pytest-asyncio==0.26.0
ruff==0.11.5
mypy==1.15.0
